Move updated Vue 2 dependencies to local third party folder
