+++ src/components/module-param-editor/module-param-editor.vue
@@ -153,7 +153,7 @@
 
 <script>
 import { mapState, mapGetters, mapMutations } from "vuex";
-import ToggleButton from "@/components/forms/vue-js-toggle-button/ToggleButton.vue";
+import ToggleButton from "@/components/third-party/vue-js-toggle-button/ToggleButton.vue";
 import EventFactory from "@/model/factories/event-factory";
 import KeyboardService from "@/services/keyboard-service";
 import ModuleParamHandler from "@/services/keyboard/module-param-handler";
